sadsadd

+++ database/migrations/2023_11_13_180251_create_table_usagers.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string("email");
             $table->string("password");
             $table->date("birthDate");
-            $table->int("role");
+            $table->unsignedBigInteger("role");
             $table->timestamps();
         });
     }
